fix(cve): remove spring boot starter web
Remove unused Spring Boot Starter Web,
include Jackson Databind instead, with
a configuration which is supposed to be
similar to the one coming from SBSW
autoconfiguration.

It will stop running a Tomcat server on
bot startup (which is absolutely unused),
reduce the number of reported CVEs
by dependabot, and reduce the bot jar size.