build: pass secrets from env vars instead of configuration file override
To pass secrets (db url, db credentials, bot token) we were required
to override the app configuration file (`application.properties`)
by copying the default one from the app resources folder to the
correct path to override it, change those secrets and add a
non-documented property `bot.token` to it.
Now, we pass secrets through environment variables and we are
no more required to override the configuration file for it.

Note: the correct path was `/application.properties` for `bootRun`
and `data/bot/application.properties` for `docker-compose.yml`

Changes:
- Rewrite `docker-compose.yml` to simplify using it locally
- Pass database configuration (url, user, password) & bot
token from environment variables instead of requiring a
`application.properties` file fully overriding the default one
  - Database configuration is hardcoded in the docker compose
file (which is obviously not the exact same in production), while
`BOT_TOKEN` comes from the `.env` file as it should not be versioned