build(docker): rewrite docker image
The existing images where no longer downloadable.
If the gradle docker image and the gradle wrapper versions
were out of sync, the wrapper was downloading the correct
version which resulted in more space used.
Now we use Temurin JDK 22 for building the app then
Temurin JRE 22 for running the app.